# API/API_definition.yaml
swagger: '2.0'
info:
  title: Pasti a domicilio
  description: definition of api in "Pasti a domicilio" project
  version: "1.0.0"
# the domain of the service
host: boiling-stream-93035.herokuapp.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /
produces:
  - text/html
consumes:
  - text/plain
# the APIs
paths:
...
  /addReservation*:
    post:
      summary: add or replace a reservation in the database
      description: add a reservation, if there isn't a reservation of that user for that day. Othervise the reservation is replaced.
      parameters:
        - name: primo
          in: formData
          description: section "primo" of the reservation
          required: true
          type: string
        - name: secondo
          in: formData
          description: section "secondo" of the reservation
          required: true
          type: string
        - name: contorno
          in: formData
          description: section "contorno" of the reservation
          required: true
          type: string
        - name: dessert
          in: formData
          description: section "dessert" of the reservation
          required: true
          type: string
      responses:
        200:
          description: the reservation has been added and the home page is shown to the user
...
  /evalutate_*:
    get:
      summary: evalutation form
      description: it returns to the user a page with the evalutation form for the specified dish
      responses:
        200:
          description: it shows the evalutation form
    post:
      summary: add the evalutation in the database
      description: add an evalutation with the specified parameters and the id of the user
      parameters:
        - name: primo
          in: formData
          description: section "primo" of the reservation
          required: true
          type: string
        - name: secondo
          in: formData
          description: section "secondo" of the reservation
          required: true
          type: string
        - name: contorno
          in: formData
          description: section "contorno" of the reservation
          required: true
          type: string
        - name: dessert
          in: formData
          description: section "dessert" of the reservation
          required: true
          type: string
      responses:
        200:
          description: the evalutation has been added and the home page is shown to the user      
...